@@ -3,27 +3,34 @@ import 'package:infinite_horizons/presentation/atoms/atoms.dart';
 
 class ListTileAtom extends StatelessWidget {
   const ListTileAtom(
-    this.title,
-    this.leading, {
+    this.title, {
+    this.trailing,
+    this.leading,
     this.subtitle,
     this.translateTitle = true,
     this.translateSubtitle = true,
+    this.onTap,
   });
 
   final String title;
   final String? subtitle;
-  final Widget leading;
+  final Widget? leading;
+  final Widget? trailing;
   final bool translateTitle;
   final bool translateSubtitle;
+  final VoidCallback? onTap;
 
   @override
   Widget build(BuildContext context) {
     return ListTile(
+      contentPadding: EdgeInsets.zero,
       title: TextAtom(title, translate: translateTitle),
       subtitle: subtitle == null
           ? null
           : TextAtom(subtitle!, translate: translateSubtitle),
       leading: leading,
+      trailing: trailing,
+      onTap: onTap,
     );
   }
 }
