@@ -1,10 +1,13 @@
-import 'package:circular_countdown_timer/circular_countdown_timer.dart';
+import 'dart:math';
+
+import 'package:custom_timer/custom_timer.dart';
 import 'package:flutter/material.dart';
+import 'package:infinite_horizons/presentation/atoms/atoms.dart';
 
 class TimerAtom extends StatelessWidget {
   const TimerAtom(this.controller, this.timer, this.callback);
 
-  final CountDownController controller;
+  final CustomTimerController controller;
   final Duration timer;
   final VoidCallback callback;
 
@@ -18,25 +21,86 @@ class TimerAtom extends StatelessWidget {
     final ThemeData themeData = Theme.of(context);
     final ColorScheme colorScheme = themeData.colorScheme;
 
-    return CircularCountDownTimer(
-      duration: timer.inSeconds,
-      controller: controller,
-      width: MediaQuery.of(context).size.width / 2,
-      height: MediaQuery.of(context).size.height / 2,
-      ringColor: Colors.grey[300]!,
-      fillColor: Colors.purpleAccent[100]!,
-      strokeWidth: 20.0,
-      strokeCap: StrokeCap.round,
-      textStyle: TextStyle(
-        fontSize: 33.0,
-        color: colorScheme.onBackground,
-        fontWeight: FontWeight.bold,
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 40.0),
+      child: CustomTimer(
+        controller: controller,
+        builder: (state, time) {
+          return SizedBox(
+            width: min(
+              MediaQuery.of(context).size.width / 2,
+              MediaQuery.of(context).size.height / 2,
+            ),
+            child: AspectRatio(
+              aspectRatio: 1.0,
+              child: ValueListenableBuilder(
+                valueListenable: controller.state,
+                builder: (context, value, child) {
+                  return GestureDetector(
+                    onTap: () {
+                      if (value == CustomTimerState.counting) {
+                        controller.pause();
+                      } else {
+                        controller.start();
+                      }
+                    },
+                    child: Stack(
+                      children: [
+                        Positioned.fill(
+                          child: CircularProgressIndicator(
+                            backgroundColor: colorScheme.primaryContainer,
+                            color: value == CustomTimerState.counting
+                                ? colorScheme.primary
+                                : colorScheme.outline,
+                            value: time.duration.inMilliseconds /
+                                timer.inMilliseconds,
+                            strokeWidth: 22,
+                            strokeCap: StrokeCap.round,
+                          ),
+                        ),
+                        Align(
+                          alignment: value != CustomTimerState.counting
+                              ? FractionalOffset.topCenter
+                              : FractionalOffset.center,
+                          child: Container(
+                            padding: value != CustomTimerState.counting
+                                ? const EdgeInsets.symmetric(
+                                    vertical: 50,
+                                    horizontal: 20,
+                                  )
+                                : null,
+                            child: TextAtom(
+                              "${time.minutes}:${time.seconds}",
+                              variant: TextVariant.title,
+                              textAlign: TextAlign.left,
+                              translate: false,
+                            ),
+                          ),
+                        ),
+                        Align(
+                          alignment: FractionalOffset.bottomCenter,
+                          child: Padding(
+                            padding: const EdgeInsets.symmetric(
+                              vertical: 38,
+                              horizontal: 38,
+                            ),
+                            child: (value != CustomTimerState.counting)
+                                ? const Icon(
+                                    Icons.play_arrow_rounded,
+                                    size: 50.0,
+                                  )
+                                : const SizedBox(),
+                          ),
+                        ),
+                      ],
+                    ),
+                  );
+                },
+              ),
+            ),
+          );
+        },
       ),
-      textFormat: CountdownTextFormat.MM_SS,
-      isReverseAnimation: true,
-      isReverse: true,
-      onComplete: onComplete,
-
     );
   }
 }
