@@ -1,5 +1,4 @@
 export 'package:infinite_horizons/presentation/organisms/intro/motivation_organism.dart';
 export 'package:infinite_horizons/presentation/organisms/intro/tips_organism.dart';
 export 'package:infinite_horizons/presentation/organisms/intro/welcome_organism.dart';
-export 'package:infinite_horizons/presentation/organisms/timer_molecule.dart';
-export 'package:infinite_horizons/presentation/organisms/top_bar_molecule.dart';
+export 'package:infinite_horizons/presentation/organisms/timer_organism.dart';
